disco/requirements.txt
fal_client==0.5.6
fpdf==1.7.2
openai==1.58.1
Pillow==9.5.0
requests==2.31.0
streamlit==1.41.1